flatbuf: add error handling when dsp is not start [1/1]
PD#SWPL-25278

Problem:
If Linux boots up without loading and starting the DSP,
using FlatBuffer will turn the process into a zombie after
process exit.

Solution:
1.Create API returns NULL
2.Read/Write API return -1
3.User applicaton executes its own error handling according to the
  failure of API call

Verify:
C1
1. Load a firmware which has a bug of crash
2. hifi4rpc_client_test --flatbuf-unit and make dsp crash
3. Reload a normal firmware
4. Execute hifi4rpc_client_test --flatbuf-unit again

+++ include/aml_flatbuf_api.h
@@ -122,9 +122,9 @@ void AML_FLATBUF_Destroy(AML_FLATBUF_HANDLE hFbuf);
  *
  * @param[in] time out in micro seconds
  *
- * @return real size of the read
+ * @return real size of the read, -1 means failure of the call
  */
-size_t AML_FLATBUF_Read(AML_FLATBUF_HANDLE hFbuf, void* buf, size_t size, int msTimeout);
+int AML_FLATBUF_Read(AML_FLATBUF_HANDLE hFbuf, void* buf, size_t size, int msTimeout);
 
 /**
  * Write data to flat buffer
@@ -141,9 +141,9 @@ size_t AML_FLATBUF_Read(AML_FLATBUF_HANDLE hFbuf, void* buf, size_t size, int ms
  *
  * @param[in] time out in micro seconds
  *
- * @return real size of the write
+ * @return real size of the write, -1 means failure of the call
  */
-size_t AML_FLATBUF_Write(AML_FLATBUF_HANDLE hFbuf, const void* buf, size_t size, int msTimeout);
+int AML_FLATBUF_Write(AML_FLATBUF_HANDLE hFbuf, const void* buf, size_t size, int msTimeout);
 
 
 /**
